+++ LeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
+﻿using LeaveManagement.Application.Exceptions;
+using LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using LeaveManagement.Application.Persistence.Contracts;
 using MediatR;
 
@@ -15,6 +16,10 @@ namespace LeaveManagement.Application.Features.LeaveRequests.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
         {
             var leaveRequest = await _leaveRequestRepository.Get(request.Id);
+            if (leaveRequest == null)
+            {
+                throw new CustomNotFoundException(nameof(leaveRequest), request.Id);
+            }
 
             await _leaveRequestRepository.Delete(leaveRequest);
