+++ LeaveManagement.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using LeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
+﻿using LeaveManagement.Application.Exceptions;
+using LeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
 using LeaveManagement.Application.Persistence.Contracts;
 using MediatR;
 using System;
@@ -20,6 +21,11 @@ namespace LeaveManagement.Application.Features.LeaveTypes.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
             var leaveType = await _leaveTypeRepository.Get(request.Id);
+
+            if (leaveType == null)
+            {
+                throw new CustomNotFoundException(nameof(leaveType), request.Id);
+            }
 
             await _leaveTypeRepository.Delete(leaveType);
